requirements.txt
contractions==0.1.73
DateTime==5.5
huggingface-hub==0.22.2
hydra-core==1.3.2
nltk==3.8.1
numpy==1.26.4
omegaconf==2.3.0
regex==2024.4.28
smart-open==6.4.0
spacy==3.7.4
textblob==0.18.0.post0
tokenizers==0.19.1
torch==2.3.0
tqdm==4.66.2
transformers==4.40.1
wordninja==2.0.0
